extract HeaderContainerRight to separate file
