+++ apps/native-component-list/src/components/HeaderIconButton.tsx
@@ -1,11 +1,6 @@
 import Ionicons from '@expo/vector-icons/build/Ionicons';
-import * as React from 'react';
-import { PropsWithChildren } from 'react';
-import { StyleSheet, TouchableOpacity, TouchableOpacityProps, View, ViewProps } from 'react-native';
-
-const HeaderContainerRight = (props: PropsWithChildren<ViewProps>) => (
-  <View {...props} style={[styles.container, props.style]} />
-);
+import React from 'react';
+import { StyleSheet, TouchableOpacity, TouchableOpacityProps } from 'react-native';
 
 type Props = TouchableOpacityProps & {
   color?: string;
@@ -20,13 +15,9 @@ const HeaderIconButton = ({ color = 'blue', disabled, name, onPress, size = 24 }
 );
 
 const styles = StyleSheet.create({
-  container: {
-    paddingRight: 8,
-    flexDirection: 'row',
-  },
   iconButton: {
     paddingHorizontal: 12,
   },
 });
 
-export { HeaderContainerRight, HeaderIconButton };
+export default HeaderIconButton;
